extract FOld logic in a separate component
