+++ src/app/js/formats/istexSummary/IstexSummaryView.js
@@ -2,21 +2,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import get from 'lodash.get';
 import { StyleSheet, css } from 'aphrodite/no-important';
-import compose from 'recompose/compose';
-import translate from 'redux-polyglot/translate';
 
-import {
-    field as fieldPropTypes,
-    polyglot as polyblotPropTypes,
-} from '../../propTypes';
+import { field as fieldPropTypes } from '../../propTypes';
 import { ISTEX_API_URL } from '../../../../common/externals';
 import injectData from '../injectData';
-import IstexItem from '../istex/IstexItem';
 import classnames from 'classnames';
-import FetchFold from './FetchFold';
-import composeAsync from '../../../../common/lib/composeAsync';
-import fetch from '../../lib/fetch';
-import { parseFetchResult } from '../shared/fetchIstexData';
+import IstexYear from './IstexYear';
 
 const styles = StyleSheet.create({
     text: {
@@ -38,65 +29,9 @@ export const getYearUrl = ({ resource, field }) => {
     )})&facet=publicationDate[perYear]&size=0&output=*`;
 };
 
-export const getVolumeUrl = ({ issn, year }) => () => ({
-    url: `${ISTEX_API_URL}/?q=(${encodeURIComponent(
-        `host.issn="${issn}" AND publicationDate:"${year}"`,
-    )})&facet=host.volume[*-*:1]&size=0&output=*`,
-});
-
-export const parseVolumeData = ({ response, error }) => {
-    if (error) {
-        throw error;
-    }
-
-    return get(response, ['aggregations', 'host.volume', 'buckets'], []).map(
-        ({ key }) => key,
-    );
-};
-
-export const getVolumeData = ({ issn, year }) =>
-    composeAsync(getVolumeUrl({ issn, year }), fetch, parseVolumeData);
-
-export const getIssueUrl = ({ issn, year, volume }) => () => ({
-    url: `${ISTEX_API_URL}/?q=(${encodeURIComponent(
-        `host.issn="${issn}" AND publicationDate:"${year}" AND host.volume:"${
-            volume
-        }"`,
-    )})&facet=host.issue[*-*:1]&size=0&output=*`,
-});
-
-export const parseIssueData = ({ response, error }) => {
-    if (error) {
-        throw error;
-    }
-
-    return get(response, ['aggregations', 'host.issue', 'buckets'], []).map(
-        ({ key }) => key,
-    );
-};
-
-export const getIssueData = ({ issn, year, volume }) =>
-    composeAsync(getIssueUrl({ issn, year, volume }), fetch, parseIssueData);
-
-export const getDocumentUrl = ({ issn, year, volume, issue }) => () => ({
-    url: `${ISTEX_API_URL}/?q=(${encodeURIComponent(
-        `host.issn="${issn}" AND publicationDate:"${year}" AND host.volume:"${
-            volume
-        }" AND host.issue:"${issue}"`,
-    )})&size=10&output=*`,
-});
-
-export const getDocumentData = ({ issn, year, volume, issue }) =>
-    composeAsync(
-        getDocumentUrl({ issn, year, volume, issue }),
-        fetch,
-        parseFetchResult,
-        ({ hits }) => hits,
-    );
-
 export class IstexSummaryView extends Component {
     render() {
-        const { formatData, field, resource, p: polyglot } = this.props;
+        const { formatData, field, resource } = this.props;
 
         return (
             <ul className={classnames('istex-year', css(styles.text))}>
@@ -104,43 +39,9 @@ export class IstexSummaryView extends Component {
                     .sort((a, b) => a.keyAsString - b.keyAsString)
                     .map(({ keyAsString }) => (
                         <li key={keyAsString} className={css(styles.li)}>
-                            <FetchFold
-                                label={keyAsString}
-                                getData={getVolumeData({
-                                    issn: resource[field.name],
-                                    year: keyAsString,
-                                })}
-                                renderData={volume => (
-                                    <FetchFold
-                                        label={`${polyglot.t('volume')}: ${
-                                            volume
-                                        }`}
-                                        getData={getIssueData({
-                                            issn: resource[field.name],
-                                            year: keyAsString,
-                                            volume,
-                                        })}
-                                        renderData={issue => (
-                                            <FetchFold
-                                                label={`${polyglot.t(
-                                                    'issue',
-                                                )}: ${issue}`}
-                                                getData={getDocumentData({
-                                                    issn: resource[field.name],
-                                                    year: keyAsString,
-                                                    volume,
-                                                    issue,
-                                                })}
-                                                renderData={document => (
-                                                    <IstexItem
-                                                        key={document.id}
-                                                        {...document}
-                                                    />
-                                                )}
-                                            />
-                                        )}
-                                    />
-                                )}
+                            <IstexYear
+                                issn={resource[field.name]}
+                                year={keyAsString}
                             />
                         </li>
                     ))}
@@ -155,7 +56,6 @@ IstexSummaryView.propTypes = {
     field: fieldPropTypes.isRequired,
     formatData: PropTypes.shape({}),
     error: PropTypes.string,
-    p: polyblotPropTypes.isRequired,
 };
 
 IstexSummaryView.defaultProps = {
@@ -166,4 +66,4 @@ IstexSummaryView.defaultProps = {
     error: null,
 };
 
-export default compose(injectData(getYearUrl), translate)(IstexSummaryView);
+export default injectData(getYearUrl)(IstexSummaryView);
